ithaka-backoffice/requirements.txt
# ============================================================================
# DEPENDENCIAS DEL PROYECTO
# ============================================================================

# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Validación de datos
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Base de datos
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Utilidades
python-multipart==0.0.6
python-dotenv>=1.0.0

# Seguridad (para cuando implementen autenticación)
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Testing (opcional, para después)
# pytest>=7.4.0
# httpx>=0.24.0